Create app.py
Flask is imported and initialized as a web application. Then, a /predict route is defined that listens for POST requests. When a request is received, a previously trained predictive model is loaded (in this example, it is assumed that the model has been saved in a model.pkl file) and the prediction is performed with the input data provided in the request. Finally, the result of the prediction is returned as a response in JSON format.